Make memory protection active on Aranym by default again. First read mint.ini and *then* try to override its values (namely no_mem_prot flag). Check (very simply) if the PMMU is present on CPU, override no_mem_prot if it is not. Make the CT60 PMMU hack actually visible, before no_mem_prot was always overriden by mint.ini.
Contributed by Miro Kropacek.

+++ sys/global.h
@@ -54,7 +54,8 @@ extern BASEPAGE *_base;	/* pointer to kernel's basepage */
 
 
 extern long mcpu; /* processor we are running */
-extern long pmmu;
+
+extern short pmmu; /* flag if mmu is present */
 
 /* for proper co-processors we must consider saving their context.
  * This variable when non-zero indicates that the BIOS considers a
